+++ lucene/core/src/java/org/apache/lucene/document/LatLonPointDistanceFeatureQuery.java
@@ -406,7 +406,7 @@ final class LatLonPointDistanceFeatureQuery extends Query {
             DocIdSetBuilder.BulkAdder adder;
 
             @Override
-            public void grow(long count) {
+            public void grow(int count) {
               adder = result.grow(count);
             }
