+++ lucene/core/src/java/org/apache/lucene/document/LongDistanceFeatureQuery.java
@@ -392,7 +392,7 @@ final class LongDistanceFeatureQuery extends Query {
             DocIdSetBuilder.BulkAdder adder;
 
             @Override
-            public void grow(long count) {
+            public void grow(int count) {
               adder = result.grow(count);
             }
