+++ lucene/core/src/java/org/apache/lucene/document/XYPointInGeometryQuery.java
@@ -76,7 +76,7 @@ final class XYPointInGeometryQuery extends Query {
       DocIdSetBuilder.BulkAdder adder;
 
       @Override
-      public void grow(long count) {
+      public void grow(int count) {
         adder = result.grow(count);
       }
