+++ lucene/core/src/java/org/apache/lucene/search/PointRangeQuery.java
@@ -169,7 +169,7 @@ public abstract class PointRangeQuery extends Query {
           DocIdSetBuilder.BulkAdder adder;
 
           @Override
-          public void grow(long count) {
+          public void grow(int count) {
             adder = result.grow(count);
           }
